+++ packages/video_player/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.10.2
+
+* **Android Only** Adds optional VideoFormat used to signal what format the plugin should try.
+
 ## 0.10.1+7
 
 * Fix tests by ignoring deprecated member use.